fix operation

+++ app/src/main/java/com/tokeninfo/ui/bean/AccountBeean.java
@@ -34,11 +34,11 @@ public class AccountBeean {
         return Symbol;
     }
 
-    public double getUsd() {
+    public float getUsd() {
         return Usd;
     }
 
-    public double getTotal() {
+    public float getTotal() {
         return Total;
     }
